+++ .gitignore
@@ -1,2 +1,2 @@
-/images
 frontend-backup
+frontend2
