+++ frontend/src/components/private.jsx
